3.tool_integration/requirements.txt
google-genai
python-dotenv
numpy
wave
sounddevice
#pyaudio